+++ package.json
@@ -28,6 +28,7 @@
     "chart.js": "^2.5.0",
     "hammerjs": "^2.0.8",
     "material-icons": "^0.6.2",
+    "ng-image-slider": "^2.8.0",
     "ng2-sticky-nav": "^10.0.0",
     "rxjs": "~6.6.0",
     "tslib": "^2.0.0",
