+++ src/app/pages/login/login.component.html
@@ -1,20 +1,23 @@
-
-<div class="holder mat-elevation-z8">
-  <div class="collage w-75 p-0">
-  <mat-grid-list cols="4" rowHeight="100px" gutterSize="15px">
-    <mat-grid-tile
-        *ngFor="let tile of Tiles"
-        [colspan]="tile.cols"
-        [rowspan]="tile.rows"
-        [ngStyle]="{'background-image':'url(' + tile.image + ')'}">
-      {{tile.text}}
-    </mat-grid-tile>
-  </mat-grid-list>
-</div>
-<div class="flipper w-25 p-0">
-  <h2>My Flipper images will Go here</h2>
-</div>
+<div class="grid-container">
+  <div class="grid-item item1">1</div>
+  <div class="grid-item item2">
+    <div class="slider ">
+    <div class="mat-elevation-z8" style="width: 100%; height: 100%; text-align:center;margin:0; padding:0;">
+      <ng-image-slider #nav
+        [images]="imageObject"
+        [infinite]="true"
+        [autoSlide]="1"
+        [showArrow]="false"
+        [imageSize]="{width: '100%', height: 433}"
+        [animationSpeed]="5"
+        slideImage="1">
+      </ng-image-slider>
+    </div>
+    </div>
   </div>
+</div>
+
+
 
 
 <!-- start nav-->
